2507 2227

+++ src/controllers/SocioEventoController.js
@@ -16,9 +16,20 @@ module.exports={
     },
 
     async post(req,res){
+        //Ainda posso verificar se de fato esse socio existe apesar do jwt ou verificar se o evento existe mesmo;
         const socio_id= req.socio_id;
         const evento_id= req.params.id;
-        //Ainda posso verificar se de fato esse socio existe apesar do jwt ou verificar se o evento existe mesmo;
+        const now = new Date();
+
+        const data_evento= await connection('eventos').where('id', evento_id).select('data').first();
+
+        const vencimento= data_evento.data
+        const parts = vencimento.split('/');
+        const data = new Date(parts[2], parts[1] - 1, parts[0]);
+
+        if(now>data){
+            return res.status(401).send({message: 'Evento ja finalizado'})
+        }
 
         const response= await connection('socio_evento').where('socio_id', socio_id).andWhere('evento_id', evento_id).select('socio_id').first();
 
@@ -34,6 +45,17 @@ module.exports={
     async delete(req,res){
         const evento_id= req.params.id;
         const socio_id= req.socio_id;
+        const now = new Date();
+
+        const data_evento= await connection('eventos').where('id', evento_id).select('data').first(); //Melhorar isso
+
+        const vencimento= data_evento.data
+        const parts = vencimento.split('/');
+        const data = new Date(parts[2], parts[1] - 1, parts[0]);
+
+        if(now>data){
+            return res.status(401).send({message: 'Evento ja finalizado'})
+        }
 
         await connection('socio_evento').where('socio_id', socio_id).andWhere('evento_id', evento_id).delete();
